refactor(storage): keep persistence checks bounded
